refactor: Ranking 분리
