--- src/main/java/site/matzip/map/MapController.java
@@ -1,13 +1,0 @@
-package site.matzip.map;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MapController {
-
-    @GetMapping("/")
-    public String showMain() {
-        return "matzip/list";
-    }
-}
