+++ src/main/java/site/matzip/member/controller/MemberController.java
@@ -137,16 +137,6 @@ public class MemberController {
         return "redirect:/usr/member/myPage";
     }
 
-    @PreAuthorize("isAuthenticated()")
-    @GetMapping("/ranking")
-    public String showRanking(Model model) {
-        List<MemberRankDTO> memberRankDtoList = memberService.findAndConvertTopTenMember();
-
-        model.addAttribute("memberRankDtoList", memberRankDtoList);
-
-        return "/ranking/member";
-    }
-
     @GetMapping("/getProfile")
     @ResponseBody
     public MemberProfileDTO getProfile(@RequestParam String nickname) {
